test(ag-logger): restructure and update test suite
- Moved unit tests to `__tests__/units` and updated imports
- Added comprehensive functional tests for AgLogger, AgLoggerManager, AgLoggerConfig, and plain output
- Removed outdated end-to-end tests
- Renamed and reorganized test files for clarity

test: ag-loggerパッケージから古くなったエンドツーエンドテストを削除

- ag-loggerパッケージの基本的なプレーンロガー出力および引数のフォーマットに関するエンドツーエンドテストを削除
